[BP-454] Add features needed for JupyterLab extension (#6)
* Add a `dry-run` option to `add_kernel` function

* Add a `kernel_exists` helper function

+++ remote_ikernel/manage.py
@@ -48,6 +48,28 @@ if sys.stdout.encoding is None:
     argparse.ArgumentParser._print_message = _print_message
 
 
+def kernel_exists(kernel_name):
+    """
+    Checks if a kernel already exists.
+
+    Parameters
+    ----------
+    kernel_name : str
+        The name of the kernel to check existence of
+
+    Returns
+    -------
+    exists : boolean
+        Boolean indicating whether a kernel with this name exists
+    """
+    try:
+        _ = ks.get_kernel_spec(kernel_name)
+    except:
+        return False
+    else:
+        return True
+
+
 def delete_kernel(kernel_name):
     """
     Delete the kernel by removing the kernel.json and directory.
@@ -141,7 +163,7 @@ def show_kernel(kernel_name):
 def add_kernel(interface, name, kernel_cmd, cpus=1, pe=None, language=None,
                system=False, workdir=None, host=None, precmd=None,
                launch_args=None, tunnel_hosts=None, verbose=False,
-               launch_cmd=None, kerneldir=None):
+               launch_cmd=None, kerneldir=None, dryrun=False):
     """
     Add a kernel. Generates a kernel.json and installs it for the system or
     user.
@@ -263,16 +285,16 @@ def add_kernel(interface, name, kernel_cmd, cpus=1, pe=None, language=None,
     else:
         username = getpass.getuser()
 
+    if not dryrun:
+        # kernel.json file installation
+        with tempdir.TemporaryDirectory() as temp_dir:
+            os.chmod(temp_dir, 0o755)  # Starts off as 700, not user readable
 
-    # kernel.json file installation
-    with tempdir.TemporaryDirectory() as temp_dir:
-        os.chmod(temp_dir, 0o755)  # Starts off as 700, not user readable
+            with open(path.join(temp_dir, 'kernel.json'), 'w') as kernel_file:
+                json.dump(kernel_json, kernel_file, sort_keys=True, indent=2)
 
-        with open(path.join(temp_dir, 'kernel.json'), 'w') as kernel_file:
-            json.dump(kernel_json, kernel_file, sort_keys=True, indent=2)
-
-        ks.install_kernel_spec(temp_dir, kernel_name,
-                               user=username, replace=True)
+            ks.install_kernel_spec(temp_dir, kernel_name,
+                                user=username, replace=True)
 
     return kernel_name, " ".join(display_name)
